+++ src/app/(home)/(components)/RecentTalkRoom.tsx
@@ -35,10 +35,12 @@ interface TalkRoomProps {
 
 const RecentTalkRoom = () => {
   const { data, isLoading, error } = useQuery(
-    TalkRoomQueryOptions.getRecentTalkRooms({
+    TalkRoomQueryOptions.getTalkRooms({
+      page: 1,
       size: 4,
       order: "recent",
       search: "",
+      sortbydate: "",
     }),
   );
