+++ src/app/components/MainThemeTitle/MainThemeTitle.tsx
@@ -1,4 +1,5 @@
 import BackButton from "@/app/summary/_component/BackButton";
+import Link from "next/link";
 import { ReactNode } from "react";
 import { ThemeMain } from "../Theme/Theme";
 
@@ -6,9 +7,17 @@ interface MainThemeTitlePrps {
   className?: string;
   children: ReactNode;
   title: string;
+  url?: string;
+  query?: string;
 }
 
-const MainThemeTitle = ({ children, title, className }: MainThemeTitlePrps) => {
+const MainThemeTitle = ({
+  children,
+  title,
+  className,
+  url,
+  query,
+}: MainThemeTitlePrps) => {
   return (
     <ThemeMain.MainTheme>
       <p className="flex flex-col sm:flex-row 2xl:mt-[78px] 2xl:mb-[23px] xl:mt-16 xl:mb-5 lg:mt-10 lg:mb-4 md:mt-6 md:mb-3 sm:mt-4 sm:mb-2">
@@ -16,9 +25,16 @@ const MainThemeTitle = ({ children, title, className }: MainThemeTitlePrps) => {
           <BackButton />
         </span>
         <span className="flex items-center sm:gap-x-1.5 md:gap-x-2 lg:gap-x-2.5 xl:gap-x-2.5 2xl:gap-x-3">
-          <span className="sm:text-base md:text-lg lg:text-xl xl:text-2xl 2xl:text-3xl cursor:pointer">
-            {title}
-          </span>
+          <Link
+            href={{
+              pathname: url || "",
+              query: query && { order: query },
+            }}
+          >
+            <span className="sm:text-base md:text-lg lg:text-xl xl:text-2xl 2xl:text-3xl cursor:pointer">
+              {title}
+            </span>
+          </Link>
           <span
             className={`2xl:size-7 xl:size-6 lg:size-5 size-4 ${className}`}
           >
